[ADS] Create RCTIMAAdsManager class
